requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.44.0
apscheduler==3.10.4
httpx==0.27.0
apprise==1.8.0
python-dotenv==1.0.1
loguru==0.7.2
rich==13.7.1
